# Make `PandasDataset` faster